product-development: a major change to how we start work on epic (#54)
This proposal is the result of a bunch of discussions I've had with Victor and
Dan about what is and isn't working about the current process. In particular:

- There's a lot of backchannel discussion going on about "what should I work on
  next?" e.g., what epics have already been claimed/spoken for, does somebody
  already have a lot of context on an epic that I need to get, etc.

- A lot of successful recent projects have gone straight from strong hunch ->
  immediately start work (like Han's user research epic), so it feels weird to
  have a "above the line" backlog that sometimes gets circumvented

- Having a weird space where you finish one epic but don't know what you're
  working on next makes it hard to start thinking/planning on an individual
  level

This new proposal makes a few major changes. Most notably, getting rid of the
line, and adding assignments for epics before you start working on them.

I'm hoping this will mostly make more explicit what's already happening
"under-the-table". A lot of these things are things we've already been doing
with the DevRel epics, which I think have been pretty successful